+++ np.CallbackURLs/src/index.js
@@ -16,19 +16,20 @@
 import pluginJson from '../plugin.json'
 
 // updateSettingsData will execute whenever your plugin is installed or updated
-import { updateSettingData } from '@helpers/NPConfiguration'
+import { updateSettingData, pluginUpdated } from '@helpers/NPConfiguration'
 
 export { xCallbackWizard } from './NPXCallbackWizard' // this makes the command function available to NotePlan (see plugin.json for details)
 
-export async function onUpdateOrInstall(): Promise<void> {
+export function onUpdateOrInstall(): void {
   // this runs after the plugin is installed or updated. the following command updates the plugin's settings data
   updateSettingData(pluginJson)
 }
 
-export async function init(): Promise<void> {
+export function init(): void {
   // this runs every time the plugin starts up (any command in this plugin is run)
-  // normally, you don't need to do anything here
-  // the command-specific entrypoints (e.g. sayHello in the helloWorld.js file is where you'll do your work)
+  DataStore.installOrUpdatePluginsByID([pluginJson['plugin.id']], true, false, false).then((r) =>
+    pluginUpdated(pluginJson, r),
+  )
 }
 
 export async function onSettingsUpdated(): Promise<void> {